Began Image Saving
The functionality for image saving is incomplete. I am able to get the path when we select an image from the gallery, but not when we take a picture with the camera. I am hoping to save the path of the file in the database (rather than an image or image blob).