Started Exams
Began working on Exam functionality for within each students record. Decided to go for a parent-child table approach for the exams DB as this seems to work better within SQLite (found through research not practice).
Im going to implement the activity of exams (for each student) very much like the activities for tasks and records, with the same components, using a recycler view to display the data. I may allow users to then press on an exam to view more about it later on (and may also add this functionality to records to make it less crammed).