Completed Notifications
Notification-based reminders for exams are now functional. There is a rather large bug with the notifications; the first notification works as intended on its own, but if you try to schedule another then the original notification will instead be sent at the new time. If the first notification has already been received when the next is sent then it schedules a notification at the new time but the content is the same, original notification.