Started Gallery
Currently there is a gallery implementation that shows the photos on the users device. They can press or hold a photo and a snackbar will say which action they did to which photo, but no other gallery functionality is implemented currently.
Plan on adding long-click to select image/s and click to deselect images that are already selected.
May also add an "add" button to the toolbar to allow the user to take a new photo in their default camera app (hence adding a photo to the apps gallery).