Gallery Functionality
Decided to implement the gallery a little differently to how we were asked. It is essentially a gallery of all images on the device. Images cannot be selected to view larger than the 4 column grid.
Images can be selected by long-clicking and deselected by short-clicking. Both of these actions also show the file path as a snackbar.
The code is not yet commented properly, though that is high-priority and will hopefully be done tomorrow.